pyproject.toml
[tool.poetry]
name = "app"
version = "1.0.0"
description = "Backend for the personal website"
authors = ["Jorge Landaeta <dev.jorge2003@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
pylint = "^3.2.6"
black = "^24.4.2"
alembic = "^1.13.2"
pydantic-settings = "^2.3.4"
psycopg = "^3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
